requirements.txt
Django==2.2.16
requests==2.26.0
djoser
djangorestframework

